requirements.txt
appnope==0.1.4
asttokens==3.0.0
cffi==1.17.1
charset-normalizer==3.4.3
comm==0.2.3
cryptography==45.0.6
debugpy==1.8.16
decorator==5.2.1
executing==2.2.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pyzmq==27.0.1
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
wcwidth==0.2.13
